third_party/blink/web_tests/fast/events/window-errorevent-02-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: DumpErrorEventAndAllowDefault is not defined
This test should trigger 'window.onerror', and leave the error unhandled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
